Revert "Revert "Merged in iulian07/handle32bit (pull request #56)""
This reverts commit 00bc1edf9e392d9269e4a174c4e8b282556b1cbf.

+++ ReadHDF5.cpp
@@ -3539,7 +3539,10 @@ ErrorCode ReadHDF5::store_file_ids( Tag tag )
 
   CHECK_OPEN_HANDLES;
 
-  typedef int tag_type;
+  //typedef int tag_type;
+  typedef long tag_type;
+  // change it to be able to read much bigger files (long is 64 bits ...)
+
   tag_type* buffer = reinterpret_cast<tag_type*>(dataBuffer);
   const long buffer_size = bufferSize / sizeof(tag_type);
   for (IDMap::iterator i = idMap.begin(); i != idMap.end(); ++i) {
